Consolidate testing utility functions
